feat: serve Mathjax components locally
Instead of sourcing the scripts and fonts from the CDN,
we bundle required assets and serve them as static files.

mathjax-full is huge (approx. 30M), which is alright
because it is not imported anywhere and will not be
a part of the final bundle.

Changes to BUILD files from running 'bazel configure'.

Excluded 'pnpm-lock.yaml' from targets for 'check-yaml' pre-commit hook,
because one of the added deps uses syntax which pre-commit-hooks falsely
identifies as invalid YAML:

https://github.com/pre-commit/pre-commit-hooks/issues/984

GH #57797