lib/output: ForceTTY can be used to disable Isatty (#60640)
For tests we do not want to use the detection on os.Stdout to change
what output we capture and assert against. In particular when running
the sg tests for me under emacs's compilation-mode I get failures due to
extra escape codes.

This is a different take on ensuring tests do not have Isatty enabled.
The previous attempt used the passed in writer to determine this.
However, on further inspection of call sites we somewhat regularly
captures the output and then only show it if something went wrong. In
those cases we should preserve Isatty of stdout so the errors look
prettier.

Test Plan: go test -short ./...