search: deduplicate symbol results from zoekt (#60697)
A query like "type:symbol repo link picker" is translated into zoekt as

  (and sym:substr:"repo" sym:substr:"link" sym:substr:"picker")

If you have a symbol RepoLinkPicker Zoekt's result type will model this
as 3 matches (which is correct) but all with the same SymbolInfo. When
we convert this into Sourcegraph's symbol information we only want to
have a match per symbol. Before this change we would translate this into
3 results, internally each having a different Column the match appears
at. By the time it would reach the user via the streaming API the column
is gone and we just have symbol information, which means effectively we
had 3 duplicate results.

This commit makes what Zoekt returns match th expectations of
Sourcegraph's symbol results. Note: we have to use a slight heuristic
here where we say two symbols are identical if they have the same line
number and symbol info.

Also note the TODO I added. The Column returned by Zoekt is where inside
of the symbol we found a match. But in downstream parts of the code we
seem to treat the Start Column as the start of the symbol. This is hard
to fix now, so instead we make a note and just tackle this.

Test Plan: added a unit test. Additionally on my local devserver
searching "repo link picker type:symbol" stopped returning duplicate
results.