gitserver: downsample internal rev-list and config in honeycomb
We are projected to hit our usage limits. Looking into it config has
spiked up to be run as much as rev-parse. I believe this is because
previously we didn't capture this in our telemetry since it was an
internal only utility.

In particular here are the top 5 events counts in the last 24 hours:

| cmd          | internal | COUNT      |
|--------------+----------+------------|
| rev-parse    | false    | 59,850,752 |
| config       | false    | 50,574,096 |
| rev-list     | false    | 14,505,488 |
| ls-tree      | false    |  2,737,872 |
| for-each-ref | false    |  2,626,320 |

https://ui.honeycomb.io/sourcegraph/datasets/gitserver-exec/result/ibtDLAoeoQP

Test Plan: go test