search: Fix delete word functionality on search results page on macOS
The event handler used for clearing the filters somehow interferes with
the search input. This commit uses the `<Shortcut />` component instead
of manually handling the event. The `Shortcut` component ignores events
originating from inputs, including the search input.

I also cleaned up the shortcut rendering code a bit to reuse the
`shortcutDisplayName` function. It's not ideal that we have to specificy
the shortcut twice (once for the keybinding and once for diplay
purposes) but it is what it is.
Often keybindings are defined in the `shared/src/keyboardShortcuts` file
but since we (probably?) don't want to show this keyboard shortcut in
the general help, it's fine to keep the definition close to where its
used.