search: do not record full error for cancel or timeout
When inspecting honeycomb there are lots of very noisy errors which are
due to timeouts or cancellations. It is noisy since it is a combined
error of all the backends that didn't finish. Additionally these are not
errors, but rather expected behaviour due to how we use context to
implement match limits and timeouts.

This commit instead does not record the error field in that case, but
instead expands the "code" field to do record this. So we now have a
more useful prometheus metric as well as the code field in honeycomb.

Test Plan: go test