insights: remove config watcher (#60720)
The computed state is so cheap that we might as well just compute it
each time we want to access it. The main difference is we won't log
insights backfiller changing. However, the old code just always logged
at startup and whenever any site config changed. It also feels somewhat
low value, so I think its fine just not to log.

Additionally the old code protected writes to task.config via a mutex,
but never used that mutex for reads. So it was in fact racey.

FYI I came across this because I was looking into reducing the logspam
in our product.

Test Plan: go test