Update help router for dev to use new docs repo
It was hardcoded to 5080 because the docs were included in the monorepo.
Now that docs are in their own repo with their own web app,
they use a different port.
Currently, that is the default Next.js port 3000.
That might change, depending on what developers do in the docs repo.
Note that this change means that in order to use the new docs,
you will need to clone the sourcegraph/docs repo, and follow its README
to run it in dev mode.