search: do type:file if all patterns are "content:" (#60782)
The way "content:" field works is we set the annotation "IsAlias" on the
pattern for it. We can't actually represent in our job framework that a
subtree only searches a specific type, so in the function computeResultTypes
we try infer a global type to run. Previously we only handled the case of one
content field. We now check if all patterns have content, and then only search
content.

Note: this is actually broken, we should atleast fail to parse the query if we
can't compute the types. EG "foo content:bar" is impossible for us to search.

Test Plan: Added a unit test.