frontend: GetAggregatedCodyStats: add nil guard for events
If there are no entries in the cody stats table (as is the case if it hasn't been used yet, events will be nil.)

In this case, then we simply return the default value of the aggregated cody stats type. It seems like this is only used for bigquery analysis, so hopefully using the default value is safe?

commit-id:46095e7a