search: use repoPagerJob for diff/commit search
This is a refactor. Instead of resolving repos inside diff/commit
search, we pull the logic out and instead wrap the `commit.SearchJob`
inside a `repoPagerJob`, just like for text search or symbol search.

Why?

To support diff/commit search in Search Jobs, the job must be wrapped in
a top-level `repoPagerJob`, which we use to scope the search job.

Test plan:
CI