Search: skip context lines when converting to line matches (#60834)
After adding context lines to chunk matches, we are now returning line matches that don't contain any of the actual matches. They have an empty OffsetAndLengths field, which means no part of that line actually matched, but we are still returning the unmatched line when splitting a chunk match into line matches.