soap cleaner: guard all delete ops on whether the set is non-empty (#60798)
This is the fix for an issue @michaellzc reported in https://sourcegraph.slack.com/archives/C05DWT4ANHH/p1709240944860449 - when `db.UserExternalAccounts().Delete(...)` is called with an empty ID set (i.e. the cleanup heuristic finds nobody that needs to be demoted/deleted), the method interprets that as a request to delete _all_ SOAP users, which is definitely not what we want. This regression was introduced in https://github.com/sourcegraph/sourcegraph/pull/56866

## Test plan

A new test case courtesy of @michaellzc

---------

Co-authored-by: Michael Lin <mlzc@hey.com>
(cherry picked from commit 33a202691ada7f3932dce973634e8a88d1bea9ae)