svelte: Introduce new feature flag for incremental rollout of Svelte on S2 (#60832)
This commit adds a new feature flag, `web-next-enabled`, to have more
control over the Svelte rollout. The plan is to enable the Svelte
prototype by default on S2 as quickly as possible.
However, not all pages are fully converted and we don't to roll out
pages that lack features people are using (e.g. the refernece panel).
A separate feature flag allows us to enable some pages by default while
having the rest be opt-in.

On the Svelte side this is handled the following ways:

1. Non existing routes already trigger a page refresh. I.e. if the user
  follows a link that is not handled by SvelteKit they'll get the React
  app automatically.
  However, the repo route is very "greedy" because it basically matches
  any path and treats it as repository name.
  This commit restricts the parameter matching by excluding known
  top-level paths. I.e. these paths are now considered as "non-existing"
  by SvelteKit and trigger a reload.
2. Existing but not enabled pages. This is handled in a
  `beforeNavigation` in the root layout. When the user navigates to a
  route that exists but not enabled we force a page refresh for the
  target URL instead.

Note that (1) means that we don't really need `data-sveltekit-reload`
anymore to trigger a page refresh, but we also use this attribute to
style attributes that lead away from the prototype.
I could imagine that we remove this styling for the first rollout on S2
though, in which case we can remove the attributes too.