svelte: Update top navigation to match S2 navigation
This commit updates the top navigation to match what's available on S2.
That means it doesn't take into account license differences or the
dotcom environment, but I tried to implement in such a way that we can
add support for this easily, by making the navigation derived from data.

To make styles consolidated I didn't reuse the existing Menu.svelte
component but used the dropdown builder from @melt-ui directly.