Svelte: port most of the filter panel behavior (#60494)
This implements almost all the behavior of the filters panel in the svelte prototype. The only missing behavior I know of is the logic around preserving or resetting filters when the type changes or the query changes. There is obviously still some missing styling, but I think we need to do a more comprehensive color and spacing audit of the prototype, so I skipped most of that.