web: Trigger page refresh SvelteKit for supported/enabled routes
This commit adds a new Link component that is aware of the `web-next`
and `web-next-enabled` feature flags, and forces a page refresh when
those feature flags are set and the path is supported/enabled.

The page refresh is done by simply rendering an `<AnchorLink />`, which
bypasses React Router.

This commit also renames the feature flag from `web-next` to
`web-next-rollout` to make the purpose clearer.