svelte/ci: Add vitest target for client/web-sveltekit
This commit adds a new target to run the unit tests inside
client/web-sveltekit with vitest.

We have an existing vitest rule (macro?) but that expects the existence
of a vitest.config.ts file, which client/web-sveltekit doesn't have. The
test config is defined in `vite.config.ts`. Therefore I'm using the
vitest binary directly.