Search: avoid crashing on frontend panic (#60878)
Because of a recent regression, frontend could panic when streaming back search
matches. These panics occur in a raw goroutine, which we didn't manually
propagate. So the panics caused frontend to crash.

This PR updates the relevant places to use `conc` objects, which properly
propagate panics. I did a "best effort" based on the original issue and didn't
audit all frontend search code.