gitserver: Pass HTTPS remote URL password as env var
This change feeds git the HTTPS URL password part via an env var through an inline shell script.
Before, the password would be part of the command args, which is viewable in ps, top and so forth.
Now, the secret can no longer be copied from the output of top.

Closes sourcegraph/security-issues#244

Closes #56810

## Test plan

Adjusted existing tests, CI runs integration and E2E tests, and tried it locally.