gitserver: Pass HTTPS remote URL password as env var (#57327)
This change feeds git the HTTPS URL user:password part via env vars through an inline shell script. Before, the password would be part of the command args, which is viewable in ps, top and so forth. Now, the secret can no longer be copied from the output of top.

Thanks to Keegan for the suggestion to use an inline script and an env var!

Closes https://github.com/sourcegraph/security-issues/issues/244

Closes https://github.com/sourcegraph/sourcegraph/issues/56810

## Test plan

CI integration tests, manually verified that clones, conn checks and so forth still work locally. Also verified that in `top` the credentials no longer show up, and no local credential helper is invoked. We believe now is the best time to merge such a change, so we have the longest possible period of us dogfooding this before a release goes out.