oidc: Add option to allow only a single identity per provider
When using an external identity provider, Sourcegraph by default doesn't care how many external identities you're adding to your account.
In some cases, this can be undesirable though, for example when trying to map a Sourcegraph user to a single identity in the outside world.

Since we cannot change the behavior here silently as people might rely on it, I decided to add a new option for this.

When enabled, the following scenario will no longer be allowed:

- I sign in to Sourcegraph for the first time using an Okta account called supererik@sourcegraph.com.
- I then go to the auth provider flow a second time (accounts tab, or through a redirect) and this time I choose my Okta account erik@sourcegraph.com
- Since I am already authenticated to Sourcegraph and it appears to Sourcegraph as if I wanted to add a new new external identity, I would have gotten a second link to an Okta identity.

Now, with this new behavior, linking a second entity for the same {ClientID, ServiceType, ServiceID} tuple but with a different externalID will be forbidden.

## Test plan

Verified locally using an Okta integration and trying to link two different accounts to the same user and see it fail.
Also added a new test to cover this behavior.