svelte: Add search history mode to search input
Closes #59208

This PR adds the search history mode to the search input. It consists of
three parts:

- Refactoring some of the logic in the React implementation to share it
  with the Svelte implementation. That includes the logic around
  managing the search history.
- Adding the history button to the search input.
- Adding logic to add a successfull search to the history.

Note that I did not copty initialializing the history from event logs.
This seemed to be a temporary solution to get some initial data when
this feature was introduced (and should probably be removed from the
React implementation).