Revert "web: Trigger page refresh to load SvelteKit for supported/enabled routes" (#60982)
This reverts commit 651a8fd89a15b4269708c4ea5dad1740c8b02fcf.

Git blame view seems to be broken because of this, and it spams the console with a bunch of scary looking stuff so I'm not confident there aren't other things broken as well. The revert applies cleanly so I figure it gives us some time to take a closer look on Monday.