graphqlbackend: allow siteadmins to view roles on dotcom (#60992)
Currently, listing roles associated with a user on dotcom fails with the error `roles are not available on sourcegraph.com`. However, roles _are_ available on dotcom and you can configure them for users (e.g. Entitler, #60795), but there's no way to check what roles a user has with  the current guard.

This relaxes the restriction to allow site admins to view a user's roles.

## Test plan

Test case