searcher: use same logic for IsRegExp as zoekt
We use the same logic as indexed search (and commit search) on deciding
if something is a regex. In theory this should mean the same thing, but
in practice we have bugs so lets use consistent behaviour.

This issue was found when investigating why a query like

  file:contains.content(foo\s*bar)

would have results with index:yes but not with index:no. Turns out the
Pattern node currently doesn't set the relevant annotations. That will
be fixed in another commit. Currently our UX highlights the inner string
with regex highlighting, so right now this is the expected behaviour. We
can debate as part of keyword work whether the pattern inside content
should require /re/ escaping.

Test Plan: CI. Note I tried to add a test case for this but it was quite
somewhat difficult to:
- adjust protocol's tests to generate a pattern with no annotations.
- test consistency across zoekt and searcher.

If anyone has advice on how to test, I would appreciate it. More than
likely the best tests will come in the next PR which fixes the missing
annotations.