worker: Fix test after go 1.22 upgrade
Seems like this test was always wrong, thanks @camdencheek!
The numWorkers variable hasn't been captured before so it always ended up being == NumTestRecords.

This revealed that the tracking of job start/end times at the millisecond precision we were doing here is just not reliable enough to not flake a lot.
So I switched the approach here to tracking the actual concurrency seen.

## Test plan

Ran the test with --count=100 locally, no more flakes.