oauth: Implement OAuth 2.0 authn and authz provider for Bitbucket Server
They added support for this in version 7.20 in the beginning of 2022.
Once complete, this will bring the UX of bitbucket server on the same level that other external identity providers for perms syncing are on.

Previously, customer HAD to use an external auth provider for both Bitbucket and Sourcegraph, and pray that the usernames will be the same on both Bitbucket Server and Sourcegraph (after all normalization steps that might apply).
This included configuration for preventing username changes on the SG side to prevent impersonation, and required to create a legacy OAuth 1.0 app on the Bitbucket side so Sourcegraph could impersonate as any actor.

## Test plan

Manual testing, need to adjust tests.