search: set Regexp annotation on patterns for FileContainsPatterns
This was the root cause of a bug where searcher and zoekt treated the
pattern it received inconsistently. The way file:contains.content()
works is it always takes a regex.

Additionally we came across a case for cody context not setting inputs
as literal. This likely fixes a bug, but I did not test further. This
came up when auditing direct creations of query.Pattern.

Test Plan: added a unit test which ensures searcher uses a regex for the
pattern.