search: simplify Deduper (#61139)
Apart from the expected deduping based on `Match.Key()`, `Deduper` kept track of the deduped results but none of the call sites ever read the results. On top of that, `Deduper` modified the matches passed to it via `Deduper.Add` which, looking at the call sites, I don't think was intentional, because all call sites keep track of results themselves. The test I fix in the second commit documents that behavior.

Notes:
- I also removed NewDedupingStream which was unused.

Test plan:
- updated unit test
- CI