guardrails: fix TestSnippetAttributionReactsToSiteConfigChanges
This test fails for me when run via go test, but for some reason passes
via bazel. I have no idea why, but the reason it fails locally for me is
we don't actually try to speak to the attribution service if there are
less than 10 lines in the snippet. So this is a terrible hack to fix
that.

Test Plan: Both of these now pass

  go test ./cmd/frontend/graphqlbackend -run TestSnippetAttributionReactsToSiteConfigChanges
  bazel test //cmd/frontend/graphqlbackend:graphqlbackend_test