svelte: Update messaging and link indicators (#61038)
This commit removes the info banner and adds an experimental badge with
popover instead. I'm sure the messaging in the popover can be improved.
Maybe we'll give people a way to permanently opt-out.

This also removes the styling for external links, which, toghether with
the changes to the repo URL matching, means that the
`data-sveltekit-reload` attributes are not necessary anymore.

Because I made changes to the popover logic, namely making it possible
to show and hide the popover on hover, I used that opportunity to
migrate from the deprecated popperjs to floating-ui, which resolves some
issues with the tooltip in hovercards.
Eventually we should probably standardize on using melt-ui for popovers
too (or whatever library we will use eventually) but for now I still
find floating-ui to be easier to use than the corresponding melt-ui
components.