Fix type imports
I ran eslint-nibble on the `client/web/`, `client/branded/`,
`client/shared/` and `client/wildcard/` folders to fix the
`@typescript-eslint/consistent-type-imports` and `@typescript-eslint/no-import-type-side-effects`
warnings.

This makes running `svelte-check` in `clint/web-sveltekit/` more useful
because it doesn't report these as error anymore (it still reports other
issues that need to be dealt with separately).