svelte: Update global Window definition (#61190)
The web app uses `window.context` and `pnpm check` (which runs
`svelte-check`) currently reports errors related to accessing this
property.

This PR extends this packages own declaration to match with the one in
web.

Eventually we'll use `window.context` too but not yet. However it's
still useful to do this now to make the output of `pnpm check` less
noisy.