msp: add support for secret volume mounts (#61185)
Closes https://github.com/sourcegraph/managed-services/issues/589, which blocks https://github.com/sourcegraph/managed-services/pull/655

This allows operators to specify volumes to mount from secrets. The configuration is intentionally limited for now, for example:

```yaml
secretVolumes:
  myVolumeName:
    mountPath: "foobar.pem"
    secret: "foobar"
```

It has similar capabilities as `secretEnv`, e.g. being able to source from external secrets and automatically have the appropriate access provisioned.

## Test plan

n/a - we use this already internally with Redis certs. This just exposes the same functionality to operators