guardrails: make graphqlbackend test race free (#61218)
We switch to using the passed in conftypes.UnifiedWatchable which allows
us to have more gaurantees around when config is read after being
watched. This allows us to simplify the implementation of syncConfMock
and I believe make it correct.

Note: previously conf.Watch was run in a goroutine. However, the
implementation of conf.Watch is to run your callback once synchronously,
then in future call the callback async. Removing the extra goroutine was
needed to make extra sure we didn't have a race in the tests.

When trying to work out why this test was failing for me in the previous
commit I noticed this simplification.

Test Plan: go test -race -count=10 -run '(Snippet|Guardrails)'