symbols: propagate errors
The purpose of this PR is to communicate to the user that an unindexed
symbol query has reached an upper limit of results. For this purpose I
used the error model recommended by grpc
[here](https://grpc.io/docs/guides/error/#richer-error-model) to
propagte errors from the symbol service to frontend.

`search.SymbolResponse` already contains a field `Err string`, but I don't think we
should keep using this, because it make it hard to propagate error types
and grpc already provides the means to propagate application errors.

Notes:
Also fixed in this PR: When we switched to grpc we stopped
propagating application errors from the symbol service. I don't think
this was intentional. Currently application errors are only visible in
the logs.

Notes: