web: Fix codeintel hovercards not working on first load (#61286)
Codeintel hovercards are not appearing when the file page initially
loads. The codeintel CodeMirror extension is never created because
`getOrCreateCodeIntelAPI` never resolves.
That's because the `platform.settings` observable never completes, which
is necessary for `lastValueFrom` to work.

The simple fix would be to use `firstValueFrom`, but upon closer
inspection, `getOrCreateCodeIntelAPI` is only used by the codemirror
blob. I refactored it to avoid having to rely on platform context, which lets
us remove platform context from a bunch of places.