svelte: Inject window.context when serving the Svelte index.html page
This commit extends the server to parse SvelteKit's index.html file as a
template and inject `window.context`.
The SvelteKit app doesn't use window.context yet but this change fixes an
issue when running `sg start web-standalone` against S2 and S2 is
configured to serve the SvelteKit app.

Couple of notes:

- I originally wanted to use a single `app.html` file for both
  "integrated" and standalone development. Having two separate files is
  not ideal, but was the easier way.
- I mostly copied the logic for generating and caching the template from
  the code that serves the esbuild bundle. There didn't seem to be an
  obvious way to make that reusable.