svelte: Persist theme choice (#61322)
Until now selecting a different theme didn't persist it in temporary
settings. This commit fixes it. I first tried an approach via context
(hence the changes to setting the context) but that didn't work as
expected (but the changes are still an improvement).

I also added a small piece of JavaScript to the HTML pages to set the
selected theme as soon as possible. This does not have any visible
effect in dev mode because the style sheets are loaded lazily.