Claude 3 support and /messages API for Enterprise (#60953)
Closes #61166

This PR adds support for Claude 3 and the /messages API to the existing
anthropic provider in the Sourcegraph instance.

To ensure a smooth experience, there are a couple of edge cases that we
need to handle.

Because the URL is configurable, a customer could hard-set it to
/complete. We need to error properly in this case. Clients might not now
what is set so they can send requests in the "old" or "new" format. We
handle conversion as best as possible however for better instruction the
clients will eventually only send prompts in the /messages format. We
introduce cody API versioning for this case. Support /complete style
prompt (with trailing assistant, "holes" of no response, system prompt
in messages) when a legacy client connects

---------

Co-authored-by: Chris Warwick <christopher.warwick@sourcegraph.com>