svelte: Update sg start enterprise-sveltekit command set (#61344)
This commit updates the enterprise-sveltekit command set to work with the
current esbuild setup. It adds a new command to start the vite build
process. The build process is quite slow so this is not a good way to
develop the SvelteKit application but it's as close to the real setup as
possible.