svelte: Remove dependency on eventLogger
One reason why the development server is relatively slow is that we are
pulling in a lot of (transitive) dependecies that are not used.
Specifically the whole wildcard package is included.

I used dependency-cruiser to find out why all of wildcard is included
and eventually discovered that it was because of importing
`eventLogger`.

`eventLogger` imports `observerQuerySelector` from `dom.ts` which in
turn imports `useObservable` from `@sourcegraph/wildcard`, which of
course imports all of wildcard.

Since `eventLogger` is deprecated anyway and we'll have to implement
proper support for event logging I've removed the depdency for now.