code monitors: don't fail job if repo empty
Previously we failed a code monitor search if any of the revisions could
not be resolved. For example, this can happen if an empty repo is added
to Sourcegraph. Now we log the error and continue with the job. The log
message is shown in the Code Monitors log view.

Notes:
- The cm_trigger_jobs table had a field `log_contents: TEXT` which was
unused. However, I anticipate we might want to add more logs in the
future so `TEXT` is not the ideal type. Instead, I decided to reuse
`execution_logs: []JSON`, which was unused, too.
- I used the oportunity to drop `cm_trigger_jobs.log_contents`

Test plan:
- manual testing
  - Added an empty repo to reroduce the error locally
  - Created a search job and verified that is passes with 0 results and
  that the logs are properly displayed in the UI
- new unit test