msp/rollouts: define custom target type in Terraform (#61366)
MSP rollouts (#59956) currently requires an additional manual step to provision via a `gcloud deploy apply` using a generated configuration YAML file. This is required because at the time, the following were not available via Terraform:

1. Cloud Deploy custom target _types_: define entities in Cloud Deploy describing the existence of custom targets using custom Skaffold scripts.
2. Cloud Deploy targets, _using_ custom target types: the Terraform resource only supports native target types, not custom targets.

In a recent GCP Terraform provider release, support for 1 was added, and this change migrates the definition currently in the generated Cloud Deploy YAML file. However, 2 is not yet supported, so we can't yet remove the manual `gcloud deploy apply` step - this is tracked in https://github.com/sourcegraph/managed-services/issues/940. This PR also improves the docstrings to better indicate what we expect to change in the future.

Closes https://github.com/sourcegraph/managed-services/issues/932

## Test plan

https://github.com/sourcegraph/managed-services/pull/939